refactor: mv /lib/types files to single types file
