+++ src/app/api/accounts/route.ts
@@ -2,8 +2,7 @@ import { eq } from "drizzle-orm";
 
 import { withAuth } from "@/lib/api/middleware";
 import { MiddlewareResponse } from "@/lib/api/response";
-import { db } from "@/lib/db";
-import { monzoAccounts } from "@/lib/db/schema/monzo-schema";
+import { db, monzoAccounts } from "@/lib/db";
 import type { Account, AccountOwner, AccountType } from "@/lib/types";
 
 export const GET = withAuth<Account>(async ({ userId }) => {
