+++ src/app/api/categories/route.ts
@@ -2,9 +2,8 @@ import { and, eq } from "drizzle-orm";
 
 import { withAccount } from "@/lib/api/middleware";
 import { MiddlewareResponse } from "@/lib/api/response";
-import { db } from "@/lib/db";
-import { monzoCategories } from "@/lib/db/schema/monzo-schema";
-import type { Category } from "@/lib/types/category";
+import { db, monzoCategories } from "@/lib/db";
+import type { Category } from "@/lib/types";
 
 export const GET = withAccount<Category[]>(async ({ accountId }) => {
   const dbCategories = await db.query.monzoCategories.findMany({
