--- src/lib/types/account.ts
@@ -1,14 +1,0 @@
-import type { AccountOwner, AccountType } from "./common";
-
-export type Account = {
-  id: string;
-  created: string;
-  type: AccountType;
-  ownerType: string;
-  isFlex: boolean;
-  productType: string;
-  currency: string;
-  owners: AccountOwner[];
-  accountNumber: string;
-  sortCode: string;
-};
