--- src/lib/types/category.ts
@@ -1,7 +1,0 @@
-export type Category = {
-  id: string;
-  name: string;
-  isMonzo: boolean;
-  createdAt: string;
-  updatedAt: string;
-};
