--- src/lib/types/common.ts
@@ -1,7 +1,0 @@
-export type AccountType = "uk_retail" | "uk_retail_joint";
-
-export type AccountOwner = {
-  user_id: string;
-  preferred_name: string;
-  preferred_first_name: string;
-};
