--- src/lib/types/index.ts
@@ -1,3 +1,0 @@
-export * from "./common";
-export * from "./account";
-export * from "./transaction";
